# Enum boxing optimizations (UnitKey)