--- a/UnitsNet.Benchmark/Conversions/FromString/ParseUnitBenchmarks.cs
+++ b/UnitsNet.Benchmark/Conversions/FromString/ParseUnitBenchmarks.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Globalization;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Jobs;
 using UnitsNet.Units;
@@ -10,6 +11,9 @@ namespace UnitsNet.Benchmark.Conversions.FromString;
 [SimpleJob(RuntimeMoniker.Net80)]
 public class ParseUnitBenchmarks
 {
+    private const int NbAbbreviations = 1000;
+    
+    private static readonly CultureInfo Culture = CultureInfo.InvariantCulture;
     private readonly Random _random = new(42);
     private string[] _densityUnits;
     private string[] _massUnits;
@@ -17,37 +21,44 @@ public class ParseUnitBenchmarks
     private string[] _volumeFlowUnits;
     private string[] _volumeUnits = [];
 
-    [Params(1000)]
-    public int NbAbbreviations { get; set; }
-
     [GlobalSetup(Target = nameof(ParseMassUnit))]
     public void PrepareMassUnits()
     {
         _massUnits = _random.GetItems(["mg", "g", "kg", "lbs", "Mlbs"], NbAbbreviations);
+        // initializes the QuantityInfoLookup and the abbreviations cache
+        Mass.TryParseUnit("_invalid", Culture, out _);
     }
 
     [GlobalSetup(Target = nameof(ParseVolumeUnit))]
     public void PrepareVolumeUnits()
     {
         _volumeUnits = _random.GetItems(["ml", "l", "L", "cm³", "m³"], NbAbbreviations);
+        // initializes the QuantityInfoLookup and the abbreviations cache
+        Volume.TryParseUnit("_invalid", Culture, out _);
     }
 
     [GlobalSetup(Target = nameof(ParseDensityUnit))]
     public void PrepareDensityUnits()
     {
         _densityUnits = _random.GetRandomAbbreviations<DensityUnit>(UnitsNetSetup.Default.UnitAbbreviations, NbAbbreviations);
+        // initializes the QuantityInfoLookup and the abbreviations cache
+        Density.TryParseUnit("_invalid", Culture, out _);
     }
 
     [GlobalSetup(Target = nameof(ParsePressureUnit))]
     public void PreparePressureUnits()
     {
         _pressureUnits = _random.GetRandomAbbreviations<PressureUnit>(UnitsNetSetup.Default.UnitAbbreviations, NbAbbreviations);
+        // initializes the QuantityInfoLookup and the abbreviations cache
+        Pressure.TryParseUnit("_invalid", Culture, out _);
     }
 
     [GlobalSetup(Target = nameof(ParseVolumeFlowUnit))]
     public void PrepareVolumeFlowUnits()
     {
         _volumeFlowUnits = _random.GetRandomAbbreviations<VolumeFlowUnit>(UnitsNetSetup.Default.UnitAbbreviations, NbAbbreviations);
+        // initializes the QuantityInfoLookup and the abbreviations cache
+        VolumeFlow.TryParseUnit("_invalid", Culture, out _);
     }
 
     [Benchmark(Baseline = true)]
@@ -56,7 +67,7 @@ public class ParseUnitBenchmarks
         MassUnit unit = default;
         foreach (var unitToParse in _massUnits)
         {
-            unit = Mass.ParseUnit(unitToParse);
+            unit = Mass.ParseUnit(unitToParse, Culture);
         }
 
         return unit;
@@ -68,7 +79,7 @@ public class ParseUnitBenchmarks
         VolumeUnit unit = default;
         foreach (var unitToParse in _volumeUnits)
         {
-            unit = Volume.ParseUnit(unitToParse);
+            unit = Volume.ParseUnit(unitToParse, Culture);
         }
 
         return unit;
@@ -80,7 +91,7 @@ public class ParseUnitBenchmarks
         DensityUnit unit = default;
         foreach (var unitToParse in _densityUnits)
         {
-            unit = Density.ParseUnit(unitToParse);
+            unit = Density.ParseUnit(unitToParse, Culture);
         }
 
         return unit;
@@ -92,7 +103,7 @@ public class ParseUnitBenchmarks
         PressureUnit unit = default;
         foreach (var unitToParse in _pressureUnits)
         {
-            unit = Pressure.ParseUnit(unitToParse);
+            unit = Pressure.ParseUnit(unitToParse, Culture);
         }
 
         return unit;
@@ -104,7 +115,7 @@ public class ParseUnitBenchmarks
         VolumeFlowUnit unit = default;
         foreach (var unitToParse in _volumeFlowUnits)
         {
-            unit = VolumeFlow.ParseUnit(unitToParse);
+            unit = VolumeFlow.ParseUnit(unitToParse, Culture);
         }
 
         return unit;