`Enum` boxing optimizations:
- introduce the `UnitKey` struct as a replacement for the conversions from `TUnit` to `Enum`
- re-mapped the `UnitAbbreviationCache` using the `UnitKey` for the concurrent dictionary
- removed all `Convert.ToInt32(..)` calls
- optimized the `QuantityInfoLookup` collections
- refactored some of the `Quantity`/`UnitConverter` code (using the `UnitParser`/`QuantityInfoLookup`)
- added some tests and benchmarks covering the new modifications

Breaking changes:
- changing the `Quantity.Names` from `string[]` to `IReadOnlyCollection<string>`
- changing the `Quantity.Infos` from `QuantityInfo[]` to `IReadOnlyList<QuantityInfo>`
- marking the `UnitConverter.ConvertByAbbreviation` method that takes a `string?` for the `IFormatProvider?` as `[Obsolete]` (created an overload)
- `UnitsNetSetup`: replaced the `ICollection<QuantityInfo>` constructor parameter with `IEnumerable<QuantityInfo>`
- renamed the `QuantityInfo.ValueType` to `QuantityInfo.QuantityType` (making the old property `[Obsolete]`)
- renamed two of the parameters of `UnitConverter.ConvertByName`