@lipchev 2025-01-31
Oh, I just remembered that there are two more benchmarks that I forgot to post:

[ParseUnitBenchmarks.cs](https://github.com/lipchev/UnitsNet/blob/unit-key/UnitsNet.Benchmark/Conversions/FromString/ParseUnitBenchmarks.cs)

Before:
<html>
<body>
<!--StartFragment-->

Method | Job | Runtime | NbAbbreviations | Mean | Error | StdDev | Ratio | RatioSD | Gen0 | Gen1 | Allocated | Alloc Ratio
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
ParseMassUnit | .NET 8.0 | .NET 8.0 | 1000 | 7.061 ms | 0.0801 ms | 0.0710 ms | 1.00 | 0.01 | 640.6250 | - | 10.32 MB | 1.00
ParseVolumeUnit | .NET 8.0 | .NET 8.0 | 1000 | 14.934 ms | 0.2945 ms | 0.2610 ms | 2.12 | 0.04 | 1171.8750 | 15.6250 | 18.76 MB | 1.82
ParseDensityUnit | .NET 8.0 | .NET 8.0 | 1000 | 15.437 ms | 0.1087 ms | 0.0849 ms | 2.19 | 0.02 | 1125.0000 | - | 18.38 MB | 1.78
ParsePressureUnit | .NET 8.0 | .NET 8.0 | 1000 | 14.054 ms | 0.2711 ms | 0.2403 ms | 1.99 | 0.04 | 1031.2500 | - | 16.62 MB | 1.61
ParseVolumeFlowUnit | .NET 8.0 | .NET 8.0 | 1000 | 22.386 ms | 0.2831 ms | 0.2510 ms | 3.17 | 0.05 | 1906.2500 | 31.2500 | 30.44 MB | 2.95
  |   |   |   |   |   |   |   |   |   |   |   |  
ParseMassUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 22.200 ms | 0.2765 ms | 0.2586 ms | 1.00 | 0.02 | 2531.2500 | - | 15.34 MB | 1.00
ParseVolumeUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 46.193 ms | 0.6411 ms | 0.5354 ms | 2.08 | 0.03 | 4909.0909 | - | 29.69 MB | 1.94
ParseDensityUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 48.149 ms | 0.7768 ms | 0.6065 ms | 2.17 | 0.04 | 5200.0000 | - | 31.25 MB | 2.04
ParsePressureUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 41.647 ms | 0.2868 ms | 0.2682 ms | 1.88 | 0.02 | 4583.3333 | - | 27.62 MB | 1.80
ParseVolumeFlowUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 68.164 ms | 0.8552 ms | 0.7141 ms | 3.07 | 0.05 | 7875.0000 | 125.0000 | 47.4 MB | 3.09

<!--EndFragment-->
</body>
</html>

After:
<html>
<body>
<!--StartFragment-->

Method | Job | Runtime | NbAbbreviations | Mean | Error | StdDev | Ratio | RatioSD | Gen0 | Gen1 | Allocated | Alloc Ratio
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
ParseMassUnit | .NET 8.0 | .NET 8.0 | 1000 | 2.844 ms | 0.0495 ms | 0.0439 ms | 1.00 | 0.02 | 472.6563 | 3.9063 | 7.58 MB | 1.00
ParseVolumeUnit | .NET 8.0 | .NET 8.0 | 1000 | 5.357 ms | 0.0552 ms | 0.0517 ms | 1.88 | 0.03 | 812.5000 | 7.8125 | 12.97 MB | 1.71
ParseDensityUnit | .NET 8.0 | .NET 8.0 | 1000 | 5.072 ms | 0.0896 ms | 0.0838 ms | 1.78 | 0.04 | 796.8750 | 7.8125 | 12.82 MB | 1.69
ParsePressureUnit | .NET 8.0 | .NET 8.0 | 1000 | 4.776 ms | 0.0796 ms | 0.0781 ms | 1.68 | 0.04 | 726.5625 | 7.8125 | 11.64 MB | 1.54
ParseVolumeFlowUnit | .NET 8.0 | .NET 8.0 | 1000 | 7.572 ms | 0.1445 ms | 0.1352 ms | 2.66 | 0.06 | 1406.2500 | 31.2500 | 22.54 MB | 2.97
  |   |   |   |   |   |   |   |   |   |   |   |  
ParseMassUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 6.915 ms | 0.0752 ms | 0.0667 ms | 1.00 | 0.01 | 1390.6250 | 7.8125 | 8.36 MB | 1.00
ParseVolumeUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 14.287 ms | 0.2017 ms | 0.1886 ms | 2.07 | 0.03 | 2468.7500 | 15.6250 | 14.89 MB | 1.78
ParseDensityUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 14.133 ms | 0.1548 ms | 0.1448 ms | 2.04 | 0.03 | 2406.2500 | 15.6250 | 14.52 MB | 1.74
ParsePressureUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 12.751 ms | 0.1183 ms | 0.1049 ms | 1.84 | 0.02 | 2187.5000 | 15.6250 | 13.18 MB | 1.58
ParseVolumeFlowUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 20.057 ms | 0.1348 ms | 0.1052 ms | 2.90 | 0.03 | 4187.5000 | 93.7500 | 25.2 MB | 3.02

<!--EndFragment-->
</body>
</html>