@lipchev 2025-01-31
[TryParseInvalidUnitBenchmarks.cs](https://github.com/lipchev/UnitsNet/blob/unit-key/UnitsNet.Benchmark/Conversions/FromString/TryParseInvalidUnitBenchmarks.cs)

Before:
<html>
<body>

<!--StartFragment-->
Method | Job | Runtime | NbAbbreviations | Mean | Error | StdDev | Median | Ratio | RatioSD | Gen0 | Gen1 | Allocated | Alloc Ratio
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
TryParseMassUnit | .NET 8.0 | .NET 8.0 | 1000 | 13.23 ms | 0.077 ms | 0.069 ms | 13.22 ms | 1.00 | 0.01 | 1218.7500 | - | 19.47 MB | 1.00
TryParseVolumeUnit | .NET 8.0 | .NET 8.0 | 1000 | 27.52 ms | 0.178 ms | 0.167 ms | 27.47 ms | 2.08 | 0.02 | 2125.0000 | - | 33.99 MB | 1.75
ParseDensityUnit | .NET 8.0 | .NET 8.0 | 1000 | 29.40 ms | 0.566 ms | 0.674 ms | 28.92 ms | 2.22 | 0.05 | 2125.0000 | - | 34.4 MB | 1.77
ParsePressureUnit | .NET 8.0 | .NET 8.0 | 1000 | 26.35 ms | 0.098 ms | 0.087 ms | 26.38 ms | 1.99 | 0.01 | 1937.5000 | - | 31.08 MB | 1.60
ParseVolumeFlowUnit | .NET 8.0 | .NET 8.0 | 1000 | 40.53 ms | 0.359 ms | 0.300 ms | 40.44 ms | 3.06 | 0.03 | 3583.3333 | 83.3333 | 57.27 MB | 2.94
  |   |   |   |   |   |   |   |   |   |   |   |   |  
TryParseMassUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 42.08 ms | 0.061 ms | 0.057 ms | 42.08 ms | 1.00 | 0.00 | 4750.0000 | - | 28.81 MB | 1.00
TryParseVolumeUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 85.93 ms | 0.689 ms | 0.644 ms | 86.12 ms | 2.04 | 0.02 | 9000.0000 | - | 54.23 MB | 1.88
ParseDensityUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 91.64 ms | 0.564 ms | 0.528 ms | 91.87 ms | 2.18 | 0.01 | 9666.6667 | - | 58.57 MB | 2.03
ParsePressureUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 78.36 ms | 1.279 ms | 1.134 ms | 78.79 ms | 1.86 | 0.03 | 8571.4286 | - | 51.69 MB | 1.79
ParseVolumeFlowUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 129.88 ms | 1.342 ms | 1.256 ms | 130.44 ms | 3.09 | 0.03 | 14750.0000 | 250.0000 | 88.7 MB | 3.08

<!--EndFragment-->
</body>
</html>

After:

<html>
<body>

<!--StartFragment-->
Method | Job | Runtime | NbAbbreviations | Mean | Error | StdDev | Ratio | RatioSD | Gen0 | Gen1 | Allocated | Alloc Ratio
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
TryParseMassUnit | .NET 8.0 | .NET 8.0 | 1000 | 5.041 ms | 0.0951 ms | 0.0934 ms | 1.00 | 0.03 | 875.0000 | 7.8125 | 14 MB | 1.00
TryParseVolumeUnit | .NET 8.0 | .NET 8.0 | 1000 | 8.916 ms | 0.1764 ms | 0.2473 ms | 1.77 | 0.06 | 1421.8750 | 15.6250 | 22.72 MB | 1.62
ParseDensityUnit | .NET 8.0 | .NET 8.0 | 1000 | 9.065 ms | 0.0963 ms | 0.0804 ms | 1.80 | 0.04 | 1453.1250 | 15.6250 | 23.28 MB | 1.66
ParsePressureUnit | .NET 8.0 | .NET 8.0 | 1000 | 8.007 ms | 0.1102 ms | 0.0921 ms | 1.59 | 0.03 | 1312.5000 | 15.6250 | 21.12 MB | 1.51
ParseVolumeFlowUnit | .NET 8.0 | .NET 8.0 | 1000 | 13.138 ms | 0.2526 ms | 0.2703 ms | 2.61 | 0.07 | 2593.7500 | 78.1250 | 41.48 MB | 2.96
  |   |   |   |   |   |   |   |   |   |   |   |  
TryParseMassUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 11.563 ms | 0.0584 ms | 0.0546 ms | 1.00 | 0.01 | 2484.3750 | 15.6250 | 14.91 MB | 1.00
TryParseVolumeUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 22.660 ms | 0.2143 ms | 0.1900 ms | 1.96 | 0.02 | 4156.2500 | 31.2500 | 25.06 MB | 1.68
ParseDensityUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 22.959 ms | 0.2643 ms | 0.2472 ms | 1.99 | 0.02 | 4187.5000 | 31.2500 | 25.13 MB | 1.69
ParsePressureUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 21.010 ms | 0.2203 ms | 0.1952 ms | 1.82 | 0.02 | 3781.2500 | 31.2500 | 22.8 MB | 1.53
ParseVolumeFlowUnit | .NET Framework 4.8 | .NET Framework 4.8 | 1000 | 33.284 ms | 0.4781 ms | 0.4472 ms | 2.88 | 0.04 | 7375.0000 | 125.0000 | 44.31 MB | 2.97

<!--EndFragment-->
</body>
</html>