@lipchev 2025-02-07
@angularsen I think this is ready, I don't have anything more to add for the time being. 

There is of course a `Part II` and `Part III` of this- my _other project_ is still *several times faster* on these... 🚀