@angularsen 2025-02-08
Oh boy, I'll need some time to just read through this 😄